renamed registries to fillerItems
